requirements.txt
beautifulsoup4
pypdf
databricks-vectorsearch